+++ CHANGELOG.md
@@ -3,3 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+
+### Added
+- Glenda parser
+  [GLM-110](https://opensource.ncsa.illinois.edu/jira/browse/GLM-110)